Aextract navigation logic to its own component
